[MAZ-73] Phase 2.6: widen ShepherdId int16→int32
Change the in-Go ShepherdId type from int16 to int32. PID range stays
[MinPID=2, MaxPID=4095] — fits in either width — but the wider type
removes a future ceiling and matches the convention for process IDs.

The sweep is mostly type-driven: callers that hold a ShepherdId (or
shepherd.PID) keep the wider type internally and narrow at the wire /
linkname boundary where required. Concretely:

- kmazarin/proc/proc.go: type ShepherdId int32 (was int16).
- kmazarin/ds/static_stack.go: extend StackElement constraint to admit
  ~uint32 | ~int32 so PID free-lists instantiate on the new type.
- kmazarin/ksyscall/user_pages.go: unmapUserPages now takes
  proc.ShepherdId — drops two narrowing casts in free_pages.go and
  runshepherd.go.

Wire-format / fixed-width fields stay int16 by design:
- shared/ipc/* (userspace ABI).
- Functions with //go:linkname to assembly (asm side is fixed-width).
- proc.NotificationEvent.{Pid,ParentPid}: the kmazarin/uring_ipc.go
  size assertion pins this struct to ipc.ProcessNotification's layout.
  Widening Pid would silently break that assertion; comment in
  notification.go documents the constraint.

Build + tests green (arm64).

+++ kmazarin/ksyscall/free_pages.go
@@ -34,7 +34,7 @@ func SyscallFreePages(arg0, arg1, _, _, _, _ uint64) int64 {
 		return -1 // EPERM
 	}
 
-	unmapUserPages(baseVA, numPages, shepherd.PageTableL0PA, int16(shepherd.PID))
+	unmapUserPages(baseVA, numPages, shepherd.PageTableL0PA, shepherd.PID)
 
 	kmem.TlbiVMALLE1()
 	kmem.DsbISH()
